+++ back/src/main/java/fr/centralesupelec/thuv/mappers/AdminCourseMapper.java
@@ -69,6 +69,7 @@ public class AdminCourseMapper {
                 )
                 .setUseStudentVolume(course.getUseStudentVolume())
                 .setStudentVolumePath(course.getStudentVolumePath())
+                .setUid(course.getUid())
                 .setVisible(course.getVisible())
                 ;