+++ back/src/main/java/fr/centralesupelec/thuv/model/Path.java
@@ -1,0 +1,28 @@
+package fr.centralesupelec.thuv.model;
+
+import lombok.Data;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+
+@Entity
+@Table(name = "paths")
+@Data
+public class Path {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotNull
+    private String value;
+
+    @NotNull
+    @Column(nullable = false, columnDefinition = "BOOLEAN DEFAULT FALSE")
+    private Boolean recursive;
+
+    @NotNull
+    @Column(nullable = false, columnDefinition = "BOOLEAN DEFAULT FALSE")
+    private Boolean changeOwner;
+
+    @NotNull
+    private Integer permissions;
+}