+++ back/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Determine whether to put courseId-<courseId> or shared-pool label on containers based on if there is a node in the swarm that has the courseId label
 
+### Added 
+- Paths for which permissions will be set before container start
+
 ### Fixed
 - Handle deployments with no sessions attached